+++ bindings/go/scip/sort_test.go
@@ -77,23 +77,23 @@ func TestSortOccurrences(t *testing.T) {
 }
 
 func TestSortRanges(t *testing.T) {
-	occurrences := []*Range{
-		NewRange([]int32{2, 3, 5}),       // rank 2
-		NewRange([]int32{11, 10, 12}),    // rank 10
-		NewRange([]int32{6, 3, 5}),       // rank 4
-		NewRange([]int32{10, 4, 8}),      // rank 6
-		NewRange([]int32{10, 10, 12}),    // rank 7
-		NewRange([]int32{0, 3, 4, 5}),    // rank 0
-		NewRange([]int32{12, 1, 13, 12}), // rank 11
-		NewRange([]int32{11, 1, 3}),      // rank 8
-		NewRange([]int32{5, 3, 5}),       // rank 3
-		NewRange([]int32{10, 1, 3}),      // rank 5
-		NewRange([]int32{12, 10, 13, 3}), // rank 13
-		NewRange([]int32{11, 4, 8}),      // rank 9
-		NewRange([]int32{12, 4, 13, 8}),  // rank 12
-		NewRange([]int32{1, 3, 3, 5}),    // rank 1
+	occurrences := []Range{
+		NewRangeUnchecked([]int32{2, 3, 5}),       // rank 2
+		NewRangeUnchecked([]int32{11, 10, 12}),    // rank 10
+		NewRangeUnchecked([]int32{6, 3, 5}),       // rank 4
+		NewRangeUnchecked([]int32{10, 4, 8}),      // rank 6
+		NewRangeUnchecked([]int32{10, 10, 12}),    // rank 7
+		NewRangeUnchecked([]int32{0, 3, 4, 5}),    // rank 0
+		NewRangeUnchecked([]int32{12, 1, 13, 12}), // rank 11
+		NewRangeUnchecked([]int32{11, 1, 3}),      // rank 8
+		NewRangeUnchecked([]int32{5, 3, 5}),       // rank 3
+		NewRangeUnchecked([]int32{10, 1, 3}),      // rank 5
+		NewRangeUnchecked([]int32{12, 10, 13, 3}), // rank 13
+		NewRangeUnchecked([]int32{11, 4, 8}),      // rank 9
+		NewRangeUnchecked([]int32{12, 4, 13, 8}),  // rank 12
+		NewRangeUnchecked([]int32{1, 3, 3, 5}),    // rank 1
 	}
-	unsorted := make([]*Range, len(occurrences))
+	unsorted := make([]Range, len(occurrences))
 	copy(unsorted, occurrences)
 
 	ranges := [][]int32{}
@@ -120,27 +120,5 @@ func TestSortRanges(t *testing.T) {
 	}
 	if diff := cmp.Diff(expected, ranges); diff != "" {
 		t.Errorf("unexpected occurrence order (-want +got):\n%s", diff)
-	}
-}
-
-func TestComparePositionToRange(t *testing.T) {
-	testCases := []struct {
-		line      int32
-		character int32
-		expected  int
-	}{
-		{5, 11, 0},
-		{5, 12, 0},
-		{5, 13, -1},
-		{4, 12, +1},
-		{5, 10, +1},
-		{5, 14, -1},
-		{6, 12, -1},
-	}
-
-	for _, testCase := range testCases {
-		if cmpRes := comparePositionToRange(5, 11, 5, 13, testCase.line, testCase.character); cmpRes != testCase.expected {
-			t.Errorf("unexpected ComparePositionSCIP result for %d:%d. want=%d have=%d", testCase.line, testCase.character, testCase.expected, cmpRes)
-		}
 	}
 }